% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.flux.data.r
\name{aggregate.flux.data}
\alias{aggregate.flux.data}
\title{Create daily aggregates of select parameters in a
standard Ameriflux data file or data frame}
\usage{
\method{aggregate}{flux.data}(df)
}
\arguments{
\item{df:}{Amerflux data file or data frame}
}
\description{
Create daily aggregates of select parameters in a
standard Ameriflux data file or data frame
}
\examples{
# with defaults, outputting a data frame
df <- aggregate.flux.data(df="AMF_BR-Sa1_2002_2011_L2_GF.txt")
}
\keyword{Ameriflux,}
\keyword{data}
\keyword{sites,}

