% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ameriflux.explorer.r
\name{ameriflux.explorer}
\alias{ameriflux.explorer}
\title{Start the AmerifluxR shiny interface}
\usage{
ameriflux.explorer()
}
\arguments{
\item{...}{none}
}
\description{
Start the AmerifluxR shiny interface
}
\examples{
ameriflux.explorer()
}
\keyword{GUI}

