% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ameriflux.info.r
\name{ameriflux.info}
\alias{ameriflux.info}
\title{Grabs the ameriflux site table from the LBL site.}
\usage{
ameriflux.info(url = "http://ameriflux.lbl.gov/sites/site-list-and-pages/")
}
\arguments{
\item{url:}{Location of the Ameriflux site table
(hopefully will not change to often, default is ok for now)}

\item{path:}{location of the phantomjs binary (system specific)}
}
\description{
Grabs the ameriflux site table from the LBL site.
}
\examples{
# with defaults, outputting a data frame
df <- ameriflux.info()
[requires the rvest package for post-processing]
http://phantomjs.org/download.html
}
\keyword{Ameriflux,}
\keyword{locations,}
\keyword{scraping}
\keyword{sites,}
\keyword{web}

