% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{amf_check_site_id}
\alias{amf_check_site_id}
\title{Check valid AmeriFlux site ID}
\usage{
amf_check_site_id(x)
}
\arguments{
\item{x}{vector or scalar of characters}
}
\value{
logical vector or scalar
}
\description{
Check if the character is a valid AmeriFlux site ID (CC-Sss)
}
