% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amf_data_aval.R
\name{amf_data_aval}
\alias{amf_data_aval}
\title{Get data variable availability}
\usage{
amf_data_aval(site_set = NULL)
}
\arguments{
\item{site_set}{a vector of character specifying the target AmeriFlux Site ID (CC-Sss). If not specified, it returns all sites.}
}
\value{
A data frame of data variable availability (per year) for selected AmeriFlux sites.
\itemize{
  \item Site_ID - Six character site identifier (CC-Sss)
  \item VARIABLE - Variable name of the data included in the BASE file
  \item BASENAME - Variable base name of the data included in the BASE file.
  \item GAP_FILLED - Whether a variable is a gap-filled variable (TRUE/FALSE)
  \item Y1990 - Percentage of data availability in the year 1990 (0-1).
  \item Y1991 - Percentage of data availability in the year 1991 (0-1).
  \item Y1992 - Percentage of data availability in the year 1992 (0-1).
  \item ...
  }
}
\description{
This function obtains the data variable availability for all or selected AmeriFlux sites.
See AmeriFlux page \url{https://ameriflux.lbl.gov/data/aboutdata/data-variables/} for details about the variable naming.
}
\examples{
#' ## Not run:
# obtain the data variable availability for all sites
data_aval <- amf_data_aval()

# obtain the data variable availability for selected sites
data_aval <- amf_data_aval(site_set = c("US-CRT","US-WPT"))

## End(Not run)

}
