% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{amf_member_info}
\alias{amf_member_info}
\title{Lists all site member info}
\usage{
amf_member_info(site_id)
}
\arguments{
\item{site_id}{Ameriflux site id}
}
\value{
Ameriflux site ids
}
\description{
This function actually lists ALL site info
of which only a fraction is used during processing
}
