% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amf_metadata_aval.R
\name{amf_metadata_aval}
\alias{amf_metadata_aval}
\title{Get metadata availability}
\usage{
amf_metadata_aval(site_set = NULL, group_only = TRUE)
}
\arguments{
\item{site_set}{a vector of character specifying the target AmeriFlux Site ID
(CC-Sss). If not specified, it returns all sites.}

\item{group_only}{logical. Should it return availability for BADM variable
groups or variables?}
}
\value{
A data frame of data variable availability (per year) for selected
 AmeriFlux sites. The first column contains the full site list. The remaining
 columns contains the number of entries for a variable or a variable group,
 with column names specifying the variable or group names.
\itemize{
  \item Site_ID - Six character site identifier (CC-Sss)
  \item ...
  }
}
\description{
This function obtains the metadata (i.e., BADM) availability for
all or selected AmeriFlux sites. See AmeriFlux page
\url{https://ameriflux.lbl.gov/data/badm/} for details about the BADM.
}
\examples{
\dontrun{
# obtain the metadata availability for all sites, at variable group levels
metadata_aval <- amf_metadata_aval()

# obtain the metadata availability for selected sites, at variable levels
metadata_aval <- amf_metadata_aval(site_set = c("US-CRT","US-WPT"),
group_only = FALSE)
}
}
