% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.exists.r
\name{dir.exists}
\alias{dir.exists}
\title{Check the existence of a directory}
\usage{
dir.exists(d)
}
\arguments{
\item{path:}{path to check for existence}
}
\description{
Check the existence of a directory
}
\examples{
# outputting TRUE or FALSE
value = dir.exists(path)
}
\keyword{files,}
\keyword{management}

