% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.ameriflux.r
\name{download.ameriflux}
\alias{download.ameriflux}
\title{Grabs the ameriflux data from the ORNL site and compiles it into one
big file.}
\usage{
download.ameriflux(site = "US-Ha1", gap_fill = TRUE)
}
\arguments{
\item{site:}{Amerflux site abbreviation, can be vector of multiple sites}

\item{gaps:}{download the gap filled or non gap filled product
(default is TRUE, allow no gaps)}
}
\description{
Grabs the ameriflux data from the ORNL site and compiles it into one
big file.
}
\examples{
# with defaults, outputting a data frame
df <- download.ameriflux(site="US-Ha1",gaps=FALSE)
}
\keyword{Ameriflux,}
\keyword{data}
\keyword{sites,}

