% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{amf_data_coverage}
\alias{amf_data_coverage}
\title{Returns a list of data coverage}
\usage{
amf_data_coverage(data_product = "BASE-BADM", data_policy = "CCBY4.0")
}
\arguments{
\item{data_product}{Data product (string). Currently, only "BASE-BADM" is supported.}

\item{data_policy}{Data policy (string). Currently, "CCBY4.0" and "LEGACY" are supported.}
}
\value{
Ameriflux data coverage
}
\description{
Ameriflux data coverage statistics
}
