% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{amf_data_coverage}
\alias{amf_data_coverage}
\title{Returns a list of data coverage}
\usage{
amf_data_coverage()
}
\value{
Ameriflux data coverage
}
\description{
Ameriflux data coverage statistics
}
