% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amf_download_bif.R
\name{amf_download_bif}
\alias{amf_download_bif}
\title{Download AmeriFlux BADM data product through web service}
\usage{
amf_download_bif(
  user_id,
  user_email,
  data_policy,
  intended_use,
  intended_use_text,
  out_dir = tempdir(),
  verbose = TRUE,
  site_w_data = FALSE
)
}
\arguments{
\item{user_id}{AmeriFlux account username (character)}

\item{user_email}{AmeriFlux account user email (character)}

\item{data_policy}{"CCBY4.0" or "LEGACY". AmeriFlux data are shared under two tiers of licenses as chosen by site's PI.
See \url{https://ameriflux.lbl.gov/data/data-policy/#data-use} for data use guidelines under each license. Note: Data use policy
selected affects which sites’ data are available for download.}

\item{intended_use}{The intended use category. Currently, it needs to be one of the followings: "synthesis", "model", "remote_sensing",
"other_research", "education", or "other"}

\item{intended_use_text}{Enter a brief description of intended use. This will be recorded in the data download log and emailed to
site's PI (free text).}

\item{out_dir}{Output directory for downloaded data}

\item{verbose}{Show feedback on download progress (TRUE/FALSE)}

\item{site_w_data}{Logical, download all registered sites (FALSE) or only sites with available BASE data (TRUE)}
}
\value{
A vector of download filenames on the local drive
}
\description{
This function allows downloading all AmeriFlux sites' BADM data files in a single file.
Note: Access to AmeriFlux data requires creating an AmeriFlux account first. Register an account through the
link \url{https://ameriflux-data.lbl.gov/Pages/RequestAccount.aspx}.For details about BADM data files, see
AmeriFlux web page \url{https://ameriflux.lbl.gov/data/aboutdata/badm-data-product/}.
}
\examples{

\dontrun{
## Download all sites with BASE data, under CCBY4.0 policy
amf_download_bif(user_id = "test",
  user_email = "test@mail.com",
  data_policy = "CCBY4.0",
  intended_use = "other",
  intended_use_text = "testing download",
  out_dir = tempdir(),
  site_w_data = TRUE)

## Download all registered sites, under LEGACY policy
amf_download_bif(user_id = "test",
  user_email = "test@mail.com",
  data_policy = "LEGACY",
  intended_use = "other",
  intended_use_text = "testing download",
  out_dir = tempdir(),
  site_w_data = FALSE)

}

}
