% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amf_site_info.R
\name{amf_site_info}
\alias{amf_site_info}
\title{Get AmeriFlux site general info}
\usage{
amf_site_info(contact_info = FALSE)
}
\arguments{
\item{contact_info}{include contact info (TRUE / FALSE)}
}
\value{
A data frame containing the following columns.
See AmeriFlux BADM standard \url{https://ameriflux.lbl.gov/data/badm/}
for detailed explanation.
\itemize{
  \item SITE_ID - Six character site identifier (CC-Sss)
  \item SITE_NAME - Site name (free text)
  \item URL_AMERIFLUX - Site web site URL, maintained by AmeriFlux (URL)
  \item TOWER_BEGAN - The starting year of flux measurement (YYYY)
  \item TOWER_END - The ending year of flux measurement (YYYY), NA if still active or unspecified
  \item IGBP - Vegetation type based on the IGBP definition (free text)
  \item GRP_LOCATION.LOCATION_LAT - Latitude of the site (decimal deg ref WGS84)
  \item GRP_LOCATION.LOCATION_LONG - Longitude of the site (decimal deg ref WGS84)
  \item GRP_LOCATION.LOCATION_ELEV - Elevation of the site above sea level (m)
  \item GRP_CLIM_AVG.MAT - Long-term mean annual average air temperature (degree C)
  \item GRP_CLIM_AVG.MAP - Long-term mean annual average precipitation (mm)
  \item GRP_CLIM_AVG.CLIMATE_KOEPPEN - Koppen climate classification (free text)
  \item DATA_START - The starting year with published AmeriFlux data (YYYY)
  \item DATA_END - The ending year with published AmeriFlux data (YYYY)
  \item TEAM_MEMBER_PI1 - Site principal investigator #1 name (First/Given Last/Family)
  \item TEAM_MEMBER_PI1_EMAIL - Site principal investigator #1 email (free text)
  \item TEAM_MEMBER_PI2 - Site principal investigator #2 name (First/Given Last/Family)
  \item TEAM_MEMBER_PI2_EMAIL - Site principal investigator #2 email (free text)
  \item TEAM_MEMBER_PI3 - Site principal investigator #3 name (First/Given Last/Family)
  \item TEAM_MEMBER_PI3_EMAIL - Site principal investigator #3 email (free text)
}
}
\description{
This function obtains the latest AmeriFlux site list and
sites' general info through the AmeriFlux web service.
}
\details{
This combines the information of various other functions.
}
